Move to util module
